move ResizeHandle to reusable ui location
